requirements.txt
daas_py_common
daas_py_config
build
requests
django
djangorestframework
psycopg2
python-dotenv
setuptools
wheel
pysolr